# Ficha7/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Ficha 7 API",
        "description": "Ficha 7 API Information",
        "contact": {
            "name": "TPSI-DWB"
        },
        "servers": [
            "http://localhost:3000"
        ]
    },
    "paths": {
        "/persons": {
            "post": {
                "tags": [
                    "person"
                ],
                "summary": "Create person",
                "description": "This can only be done by the logged in user.",
                "operationId": "createPerson",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created person object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "person"
                ],
                "summary": "Update person",
                "description": "This can only be done by the logged in user.",
                "operationId": "updatePerson",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated person object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "Get persons",
                "description": "Get all persons.",
                "operationId": "getPersons",
                "produces": [
                    "application/xml",
                    "application/json"
                ],               
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No persons found"
                    }
                }
            }
        },
        "/persons/{id}": {
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "Get person by id",
                "description": "Get person by ID.",
                "operationId": "getPersonById",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id that needs to be fetched. Use id 1 for testing. ",
                        "required": true,
                        "type": "string"
                    }                    
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "person"
                ],
                "summary": "Delete person by id",
                "description": "Delete person by ID.",
                "operationId": "deletePersonById",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id that needs to be fetched. Use id 1 for testing. ",
                        "required": true,
                        "type": "string"
                    }                    
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }               
        }
    },
    "definitions": {
        "Person": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-primary-key": true
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "age": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "xml": {
                "name": "Person"
            }
        }
    }
}